+++ lib/src/signing.rs
@@ -25,6 +25,7 @@ use thiserror::Error;
 use crate::backend::CommitId;
 use crate::config::ConfigGetError;
 use crate::gpg_signing::GpgBackend;
+use crate::gpg_signing::GpgsmBackend;
 use crate::settings::UserSettings;
 use crate::ssh_signing::SshBackend;
 use crate::store::COMMIT_CACHE_CAPACITY;
@@ -179,10 +180,10 @@ impl Signer {
     pub fn from_settings(settings: &UserSettings) -> Result<Self, SignInitError> {
         let mut backends: Vec<Box<dyn SigningBackend>> = vec![
             Box::new(GpgBackend::from_settings(settings).map_err(SignInitError::BackendConfig)?),
+            Box::new(GpgsmBackend::from_settings(settings).map_err(SignInitError::BackendConfig)?),
             Box::new(SshBackend::from_settings(settings).map_err(SignInitError::BackendConfig)?),
             #[cfg(feature = "testing")]
             Box::new(TestSigningBackend),
-            // Box::new(X509Backend::from_settings(settings).map_err(..)?),
         ];
 
         let main_backend = settings