+++ CHANGELOG.md
@@ -43,6 +43,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* `jj sign` can now sign with PKCS#12 certificates through the `gpgsm` backend.
+
 * `jj sign` will automatically use the gpg key associated with the author's email
   in the absence of a `signing.key` configuration.